requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==32.1.0
psycopg2==2.9.9
PyJWT==2.9.0
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
